pyproject.toml
[tool.poetry]
name = "rerun_nova"
version = "0.0.1"
description = ""
readme = "README.md"
package-mode = false

include = ["rerun_nova"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
rerun-sdk = "^0.21.0"
wandelbots-api-client = "^24.8.1"
python-decouple = "^3.8"
requests = "^2.32.3"
APScheduler = "^3.11.0"
scipy = "^1.14.1"
trimesh = "^4.5.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
